[nemo-qml-plugin-messages] Fix a possible assertation when creating channels
The client handler can receive the channel created by a channel request
before the channelRequestSucceeded signal arrives, in which case it will
call setChannel directly (done primarily for un-requested channels).

Fixed those code paths to have the same behavior and not trigger an
assertation on the state of the request.