[nemo-qml-plugin-messages] Only register as a Telepathy handler if explicitly given a handler name
This allows clients to register under a handler name of their choice,
and allows creation of a TelepathyChannelManager instance without
registering as a Telepathy handler (which is useful for clients that
wish to establish channels to send messages without intending to receive
or display any).

Note that text channels may be discarded by Telepathy if there is not at
least one registered handler on the system; generally, there should be a
primary messaging application that does so.