[nemo-qml-plugin-messages] Separate ChannelManager and Tp ClientHandler
This fixes a crash on destruction of ChannelManager.

Tp::ClientRegistrar requires a shared pointer to the
Tp::AbstractClientHandler subclass being registered; the QML-managed
ChannelManager object does not mesh well with telepathy's shared
pointers, which resulted in the object being deleted twice.

The ClientHandler subclass is now a separate object, so it can be
managed exclusively by Tp::SharedPtr.